+++ data/src/main/java/com/example/data/di/RepositoryModule.kt
@@ -1,7 +1,7 @@
 package com.example.data.di
 
-import com.example.data.remote.repository.BookRepositoryImpl
-import com.example.data.remote.source.BookDataSource
+import com.example.data.repository.BookRepositoryImpl
+import com.example.data.source.remote.BookDataSource
 import com.example.domain.repository.BookRepository
 import dagger.Module
 import dagger.Provides
