+++ data/src/main/java/com/example/data/network/NetworkInterceptor.kt
@@ -1,4 +1,4 @@
-package com.example.data.remote.network
+package com.example.data.network
 
 import okhttp3.Interceptor
 import okhttp3.Response
@@ -7,8 +7,8 @@ import javax.inject.Inject
 class NetworkInterceptor @Inject constructor(private val networkConfig: NetworkConfig) : Interceptor {
     override fun intercept(chain: Interceptor.Chain): Response {
         val request = chain.request().newBuilder()
-            .addHeader("X-Naver-Client-Id", networkConfig.clientId)
-            .addHeader("X-Naver-Client-Secret", networkConfig.clientSecret)
+            .addHeader("X-Naver-Client-Id", NetworkConfig.clientId)
+            .addHeader("X-Naver-Client-Secret", NetworkConfig.clientSecret)
             .build()
         return chain.proceed(request)
     }
