+++ presentation/src/main/java/com/example/presentation/viewmodel/SearchViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.domain.model.Book
-import com.example.domain.usecase.GetSearchBookUseCase
+import com.example.domain.usecase.search.GetSearchBookUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
